i end validation in student and started to validte subject and near to complete the remain validation is delete 12_9_2025

+++ SchoolManagmentSys/Controllers/StudentsController.cs
@@ -221,11 +221,26 @@ namespace API.Controllers
         public IActionResult Delete(long id)
         {
             var student = _studentRepository.Find(x => x.Id == id).FirstOrDefault();
-            if (student == null)
-                return NotFound();
+            bool hadEnrolledSub = _studentRepository.Find(x => x.Id==x.StudentSubjects.FirstOrDefault(ss=>ss.StudentId==student.Id).StudentId, x => x.StudentSubjects).Any();
 
-            _studentRepository.Delete(student);
-            return Ok();
+
+
+
+
+            if (student == null)
+            {
+                return NotFound();
+            }
+            if (hadEnrolledSub)
+            {
+                return Conflict("Cannot delete student with enrolled subjects.");
+            }
+            else
+            {
+
+                _studentRepository.Delete(student);
+                return Ok("Student Deleted Successfully");
+            }
         }
